+++ earth2studio/lexicon/__init__.py
@@ -24,5 +24,6 @@ from .goes import GOESLexicon
 from .hrrr import HRRRFXLexicon, HRRRLexicon
 from .ifs import IFSLexicon
 from .imerg import IMERGLexicon
+from .jpss import JPSSLexicon
 from .ncar import NCAR_ERA5Lexicon
 from .wb2 import WB2ClimatetologyLexicon, WB2Lexicon