+++ earth2studio/models/dx/corrdiff.py
@@ -82,24 +82,28 @@ class CorrDiff(torch.nn.Module, AutoModelMixin):
         Core pytorch model for diffusion step
     regression_model : torch.nn.Module
         Core pytorch model for regression step
-    lat_grid : torch.Tensor
-        Output latitude grid tensor
-    lon_grid : torch.Tensor
-        Output longitude grid tensor
+    lat_input_grid : torch.Tensor
+        Input latitude grid of size [in_lat, in_lon]
+    lon_input_grid : torch.Tensor
+        Input longitude grid of size [in_lat, in_lon]
+    lat_output_grid : torch.Tensor
+        Output latitude grid of size [out_lat, out_lon]
+    lon_output_grid : torch.Tensor
+        Output latitude grid of size [out_lat, out_lon]
     in_center : torch.Tensor
-        Model input center normalization tensor
+        Model input center normalization tensor of size [in_var]
     in_scale : torch.Tensor
-        Model input scale normalization tensor
-    invariant_center : torch.Tensor
-        Invariant features center normalization tensor
-    invariant_scale : torch.Tensor
-        Invariant features scale normalization tensor
+        Model input scale normalization tensor of size [in_var]
     out_center : torch.Tensor
-        Model output center normalization tensor
+        Model output center normalization tensor of size [out_var]
     out_scale : torch.Tensor
-        Model output scale normalization tensor
+        Model output scale normalization tensor of size [out_var]
     invariants : OrderedDict | None, optional
         Dictionary of invariant features, by default None
+    invariant_center : torch.Tensor | None, optional
+        Model invariant center normalization tensor of size [len(invariants)], by default None
+    invariant_scale : torch.Tensor | None, optional
+        Model invariant scale normalization tensor of size [len(invariants)], by default None
     number_of_samples : int, optional
         Number of high resolution samples to draw from diffusion model, by default 1
     number_of_steps : int, optional